// src/app/Models/crime.model.ts
export class Crime {
    id: number;
    name: string;
    gender: string;
    age: string;
    email?: string;
    phoneNumber?: string;
    contactPreference: string;
    caseNumber: string;
    fileNumber: string;
    complaint: string;
    investigationAssignee: string;
    progressDetail: string;
    medium?: string;
    attachment?: string;
    relatedLink?: string;
    date: Date;
    remarks?: string;
    isActive: boolean;
    nationality?: string;
    photoPath?: string;
}